changed header change listener so it return header id along with the header rect, so it's easier to consume on the client side

+++ library/src/main/java/com/timehop/stickyheadersrecyclerview/StickyRecyclerHeadersAdapter.java
@@ -12,7 +12,7 @@ public interface StickyRecyclerHeadersAdapter<VH extends RecyclerView.ViewHolder
    * @param position
    * @return
    */
-  public long getHeaderId(int position);
+  public int getHeaderId(int position);
 
   /**
    * Creates a new ViewHolder for a header.  This works the same way onCreateViewHolder in
